// src/app/Shared/Models/Client.ts
export class Client {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  confirmPassword: string;
  country: string;
  contact1: string;
  username: string;
  contact2: string;
  role: string;
}
export class ClientRegistered {
  ID: number;
  country: string;
  contact1: string;
  Username: string;
  role: string;
  email: string;
  name: string;
}

export class Logged {
  id: number;
  username: string;
  role: string;
}